@@ -1,6 +1,6 @@
 package org.rembx.jeeshop.catalog.model;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.rembx.jeeshop.catalog.model.Discount.ApplicableTo.ORDER;
