@@ -1,8 +1,7 @@
 package org.rembx.jeeshop.order;
 
-import org.junit.Test;
 
-import static org.junit.Assert.*;
+import org.junit.jupiter.api.Test;
 
 // TODO!
 public class EligibleDiscountsCT {
