@@ -1,6 +1,6 @@
 package org.rembx.jeeshop.order;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
