@@ -1,6 +1,7 @@
 package org.rembx.jeeshop.order.model;
 
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import java.text.SimpleDateFormat;
 
