@@ -1,6 +1,8 @@
 package org.rembx.jeeshop.catalog;
 
 
+import io.quarkus.hibernate.orm.PersistenceUnit;
+import io.quarkus.undertow.runtime.HttpSessionContext;
 import org.rembx.jeeshop.catalog.model.Catalog;
 import org.rembx.jeeshop.catalog.model.CatalogPersistenceUnit;
 import org.rembx.jeeshop.catalog.model.Category;
@@ -10,15 +12,17 @@ import org.rembx.jeeshop.rest.WebApplicationException;
 import javax.annotation.Resource;
 import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
-import javax.ejb.SessionContext;
-import javax.ejb.Stateless;
+import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.transaction.Transactional;
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
@@ -34,26 +38,25 @@ import static org.rembx.jeeshop.role.JeeshopRoles.ADMIN_READONLY;
  */
 
 @Path("/catalogs")
-@Stateless
+@Transactional
+@RequestScoped
 public class Catalogs {
 
-    @PersistenceContext(unitName = CatalogPersistenceUnit.NAME)
-    private EntityManager entityManager;
+    @PersistenceUnit(CatalogPersistenceUnit.NAME)
+    EntityManager entityManager;
 
-    @Inject
     PresentationResource presentationResource;
 
-    @Inject
     private CatalogItemFinder catalogItemFinder;
 
-    @Resource
-    private SessionContext sessionContext;
+    @Context
+    SecurityContext sessionContext;
 
     public Catalogs() {
 
     }
 
-    public Catalogs(EntityManager entityManager, CatalogItemFinder catalogItemFinder) {
+    Catalogs(EntityManager entityManager, CatalogItemFinder catalogItemFinder) {
         this.entityManager = entityManager;
         this.catalogItemFinder = catalogItemFinder;
     }
