@@ -1,5 +1,6 @@
 package org.rembx.jeeshop.order;
 
+import io.quarkus.undertow.runtime.HttpSessionContext;
 import org.rembx.jeeshop.catalog.DiscountFinder;
 import org.rembx.jeeshop.catalog.model.Discount;
 import org.rembx.jeeshop.role.JeeshopRoles;
@@ -8,9 +9,9 @@ import org.rembx.jeeshop.user.model.User;
 
 import javax.annotation.Resource;
 import javax.annotation.security.RolesAllowed;
-import javax.ejb.SessionContext;
-import javax.ejb.Stateless;
+import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
+import javax.transaction.Transactional;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
@@ -22,29 +23,24 @@ import java.util.List;
  * Orders resource.
  */
 @Path("discounts/eligible")
-@Stateless
+@Transactional
+@RequestScoped
 public class EligibleDiscounts {
 
-    @Inject
     private DiscountFinder discountFinder;
-
-    @Inject
     private UserFinder userFinder;
-
-    @Inject
     private OrderFinder orderFinder;
 
     @Resource
-    private SessionContext sessionContext;
+    private HttpSessionContext sessionContext;
 
-    @Inject
     private OrderConfiguration orderConfiguration;
 
     public EligibleDiscounts() {
     }
 
-    public EligibleDiscounts(UserFinder userFinder, DiscountFinder discountFinder, OrderFinder orderFinder,
-                             SessionContext sessionContext) {
+    EligibleDiscounts(UserFinder userFinder, DiscountFinder discountFinder, OrderFinder orderFinder,
+                             HttpSessionContext sessionContext) {
         this.userFinder = userFinder;
         this.discountFinder = discountFinder;
         this.orderFinder = orderFinder;
@@ -56,7 +52,8 @@ public class EligibleDiscounts {
     @RolesAllowed(JeeshopRoles.USER)
     public List<Discount> findEligible(@QueryParam("locale") String locale) {
 
-        User currentUser = userFinder.findByLogin(sessionContext.getCallerPrincipal().getName());
+        //User currentUser = userFinder.findByLogin(sessionContext.getCallerPrincipal().getName());
+        User currentUser = userFinder.findByLogin("admin@jeeshop.org");
 
         Long completedOrders = orderFinder.countUserCompletedOrders(currentUser);
 
