@@ -1,8 +1,9 @@
 package org.rembx.jeeshop.order;
 
 import javax.annotation.security.PermitAll;
-import javax.ejb.Stateless;
+import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
+import javax.transaction.Transactional;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
@@ -12,16 +13,16 @@ import javax.ws.rs.core.MediaType;
  * Orders resource.
  */
 @Path("fees")
-@Stateless
+@Transactional
+@RequestScoped
 public class Fees {
 
-    @Inject
     private OrderConfiguration orderConfiguration;
 
     public Fees() {
     }
 
-    public Fees(OrderConfiguration orderConfiguration) {
+    Fees(OrderConfiguration orderConfiguration) {
         this.orderConfiguration = orderConfiguration;
     }
 
