@@ -1,12 +1,12 @@
 package org.rembx.jeeshop.user;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import io.quarkus.hibernate.orm.PersistenceUnit;
 import org.rembx.jeeshop.user.model.Role;
 import org.rembx.jeeshop.user.model.RoleName;
 import org.rembx.jeeshop.user.model.UserPersistenceUnit;
 
 import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 
 import static org.rembx.jeeshop.user.model.QRole.role;
 
@@ -15,8 +15,8 @@ import static org.rembx.jeeshop.user.model.QRole.role;
  */
 public class RoleFinder {
 
-    @PersistenceContext(unitName = UserPersistenceUnit.NAME)
-    private EntityManager entityManager;
+    @PersistenceUnit(UserPersistenceUnit.NAME)
+    EntityManager entityManager;
 
     public RoleFinder() {
     }
